create EventUtils.js + move modals over to page folder if page-specific
