+++ src/components/DisplayEvents.css
@@ -45,8 +45,6 @@
   margin-top: 12px;
 }
 
-/* Event Sizing */
-
 .event.small,
 .event.medium {
   display: flex;
