+++ src/components/DisplayEvents.jsx
@@ -7,7 +7,6 @@ export default function DisplayEvents({ events }) {
   const hourGridLines = new Array(24).fill(null);
   const { setModalContext } = useModalContext();
 
-  // always have up to date copy of events for that day
   return (
     <div className="calendar-column">
       {hourGridLines.map((_, i) => (
