+++ src/components/HoursList.css
@@ -1,6 +1,4 @@
 .hours {
-  /* border-right: 1px solid gray; */
-  /* transform: translateY(-8px); */
   display: inline-flex;
   flex-direction: column;
 }
